requirements.txt
discord-cooldown>=0.1.5
python-dotenv>=1.0.0
pycolorise==1.0.0
